test: cleanup duplicate codes
